Rename gas fees file
